Release 5.4.1
There is no 5.4 (well there is but it's partially broken as the
vendoring is backport.shutils-get-terminal-size was not done correctly so break
some stuf).

And we keep above mentioned package as a dependency as we try to import
it, but as backport packages are a mess, we fallback on a vendored
version if the import fail. SO no change to the recipe